+++ xdef/changelog.md
@@ -1,9 +1,10 @@
 # Version ${version}, release-date ${release.date}
 
-# Version 42.2.25, release-date 2025-03-27
+# Version 42.2.25, release-date 2025-03-28
 * Fixed error when parsing email address when the local part contains several consecutive hypen characters.
-* The `boolean isEmoji()` method has been added to the `org.xdef.sys.SParser` interface, which returns
-  the value of the character if it is an emoji and moves to the next character or returns the value `NOCHAR`.
+* A new method `boolean isEmoji()` has been added to the `org.xdef.sys.SParser` interface. If the current
+  character is an emoji, the method will return the value of that character and move to the next character,
+  otherwise it will only return `NOCHAR`.
 
 # Version 42.2.24, release-date 2025-03-25
 * Fixed error when parsing email address when the domain part contains several consecutive hypen characters.